pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rd"
version = "0.1.0"
description = "ReliableData — a SQLite3-backed key/value store with a REST interface"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "fastapi>=0.110.0",
    "pycryptodome>=3.20",
    "rich>=13.0",
    "sqlalchemy>=2.0",
    "uvicorn[standard]>=0.29.0",
]

[project.scripts]
rd = "rd.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/rd"]

[tool.uv]

[dependency-groups]
dev = [
    "approvaltests>=14.7",
    "black>=24.0",
    "coverage[toml]>=7.6",
    "flake8>=7.0",
    "httpx>=0.27.0",
    "isort>=5.13",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[tool.coverage.run]
branch = true
source = ["src/rd"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
